Fixed some small issues with updating assigned_to, created_by and last_edited_by for the epic and story forms, made epics, stories, tracking columns, teams and users deletable and finished help section (issues #40 and #36)

+++ src/Requirements_and_value_identification_tool/react-app/src/components/tracking_column_forms/edit_tracking_column_form.js
@@ -123,10 +123,14 @@ class EditColumnForm extends Component {
     deleteColumn = () => {
         //move stories to undefined
         this.redefine(this.state.stories, 'undefined');
-        axios.delete(API_URL_TRACKING_COLUMN_DETAILS + this.state.column_id).then(() => {
-            this.props.resetState();
-            this.props.toggle();
-        });
+
+        if (window.confirm("Delete tracking column:  '" + this.state.title + "'?")) {
+            axios.delete(API_URL_TRACKING_COLUMN_DETAILS + this.state.column_id).then(() => {
+                this.props.resetState(this.state);
+                this.props.toggle();
+                alert('Column deleted');
+            });
+        }
     };
 
     onStoryAddition = (e) => {
@@ -297,7 +301,7 @@ class EditColumnForm extends Component {
 
                 <Button className="btn-primary">Update Column</Button>
 
-                <Button className="btn-danger float-end" onClick={this.deleteColumn}>
+                <Button className="btn-danger float-end" onClick={() => this.deleteColumn()}>
                     Delete Column
                 </Button>
             </Form>
